Even more sounds
Death sounds don't work, they get unloaded from the stats sheet on scene load
