+++ stc_game/Assets/Prefabs/Weapons/CrystalSwordComplete 1.prefab
@@ -107,7 +107,7 @@ GameObject:
   - component: {fileID: 7063556923144885398}
   - component: {fileID: 7018611949352354718}
   m_Layer: 2
-  m_Name: CrystalSwordComplete
+  m_Name: CrystalSwordComplete 1
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -200,6 +200,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a04e8789faa6c4c4f8c090d20a9e9278, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  sound: {fileID: 8300000, guid: 4ee1f9f5a0580a3488c0176cb02b123d, type: 3}
+  volume: 0.1
   hittingTarget: 0
   target: {fileID: 0}
 --- !u!114 &7063556923144885398
