+++ stc_game/Assets/Prefabs/Weapons/MagicStaff.prefab
@@ -14,7 +14,7 @@ GameObject:
   - component: {fileID: 5337873147915807779}
   - component: {fileID: 2199265197987293831}
   m_Layer: 0
-  m_Name: SM_Prop_WizardStaff_02
+  m_Name: MagicStaff
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -106,6 +106,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a04e8789faa6c4c4f8c090d20a9e9278, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  sound: {fileID: 8300000, guid: 4ee1f9f5a0580a3488c0176cb02b123d, type: 3}
+  volume: 0.1
   hittingTarget: 0
   target: {fileID: 0}
 --- !u!1 &4251138641213861744
