+++ stc_game/Assets/Scripts/CharacterBehavior/Drops/Gold.cs
@@ -1,11 +1,6 @@
 ﻿using Devdog.General;
 using Devdog.QuestSystemPro;
-using System;
 using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using UnityEngine;
 
 namespace Assets.Scripts.CharacterBehavior.Drops
