--- stc_game/Assets/Scripts/Combat/Abilities/ParticleActionScripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cf4b978c78807354abee1b579e75c5c8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
