+++ stc_game/Assets/Scripts/Combat/Abilities/ScriptableObjects/Sounds/SprintSound.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 26d642fe547622643b4ab630938e1538
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
