+++ stc_game/Assets/Scripts/Combat/Abilities/SoundActionScript/ParticleActionScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cf4b978c78807354abee1b579e75c5c8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
