+++ stc_game/Assets/Scripts/Combat/Abilities/SoundActionScript/ParticleActionScripts/InstantParticle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2202453ce52f6a143a3529f6e3fb5e23
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
