+++ stc_game/Assets/Scripts/Combat/Abilities/SoundActionScript/SoundAction.cs
@@ -1,7 +1,6 @@
 ﻿using Assets.Scripts.CharacterBehavior.BaseClasses;
 using Assets.Scripts.CharacterBehavior.Combat;
 using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 [CreateAssetMenu(menuName = "Ability/AbilityAction/SoundAction/PlaySound")]
