+++ stc_game/Assets/Sounds/Death.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 47b298229160cc442a235ae8f3d88a22
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
