+++ stc_game/Assets/Sounds/Music/ebunny_-_Enchanted_Violin.mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: d56223fb625c58d4a8e26448cb242229
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
