+++ stc_game/Assets/Sounds/hits/zapsplat_warfare_sword_swing_fast_whoosh_metal_004.mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 3908eb9a6471f1c40bea2f9a306e38e5
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
